requirements.txt
discord.py[voice]>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
PyNaCl>=1.5.0
pytz>=2024.1
SQLAlchemy>=2.0
aiosqlite

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
flake8>=6.0.0

# Optional but recommended
# asyncpg>=0.29.0  # For PostgreSQL database

# Music (disabled for now)
# wavelink>=3.4.1